Allow deep-frozen Array and Hash as shareable (naive implementation) (#12)
To allow an access to constants that contain deep-frozen Array and Hash.

Note that the implementation is just temporal.  It has two problems:

(1) The object is traversed for every check, which may be very slow.
(2) It uses recursion, which may cause stack overflow.

Still, I think this is useful for experiment of Ractor.